wireless: Enable WIPHY suspend, with a twist
Don't disconnect from the AP at suspend time since we want to keep
the possibility of pinging the device while it's suspended.
To do this leave the device connected to the AP and let qcacld's
suspend/resume functionality trigger an interrupt on incoming data.
Do this without requiring the interface to be wowlan-enabled.