Partially revert "ARM64: dts: Switch mt6765 to 2-stage-init"
Halium's mount-android-partitions searches the DT fstab for additional
partitions to be mounted in the container. It just so happens that Huawei
used the /odm partition, which isn't mounted by default in the script, nor
does it appear the fstab (as this device uses DT fstab).

Revert this change to bring back the DT fstab entries to ensure /odm
gets mounted inside the container, but only include /vendor and /odm.

This partially reverts commit c61009aa7294ff62ab43c27736a4c8b42261d52e.

Change-Id: I8fda7628d8a2d1ed0c5d2a5ed7c2ce9885e9685c
Signed-off-by: bengris32 <bengris32@protonmail.ch>